requirements.txt
psutil
pywin32
pyttsx3
SpeechRecognition
PyAudio
pvporcupine
python-dotenv
fuzzywuzzy
python-Levenshtein
pyautogui
customtkinter
